[fix] DB, heroku
