+++ .env
@@ -1,2 +1,4 @@
 FLASK_APP=run.py
 FLASK_DEBUG=TRUE
+
+REDIS_URL=redis://h:pbeaf42ca43c79d531fcc0b97880ecdb8dab7e84c173f2de68179ae408b770886@ec2-52-45-40-253.compute-1.amazonaws.com:27019
