+++ mods/DB.py
@@ -4,25 +4,13 @@ from mods.read_config import read_config
 
 class ReadWriteDB:
     def __init__(self):
-        """ local版
-        DBMS = os.getenv("DBMS")
-        USER = os.getenv("USER")
-        PASS = os.getenv("PASS")
-        HOST = os.getenv("HOST")
-        DB = os.getenv("DB")
-        TABLE = os.getenv("TABLE")
-
-        db = dataset.connect("{0}://{1}@{2}/{3}".format(DBMS, USER, HOST, DB))
-        """
-
         ini = read_config("config")
         read_ini = ini["DATABASE"]
         URI = read_ini.get("URI")
         TABLE = read_ini.get("TABLE")
         
-        db = dataset.connect(URI)
-        
-        self.table = db[TABLE]
+        self.__db = dataset.connect(URI)
+        self.table = self.__db[TABLE]
 
     def Write(self, data):
         self.table.insert(data)
@@ -32,3 +20,6 @@ class ReadWriteDB:
 
     def DeleteByID(self, quiz_id):
         self.table.delete(id = quiz_id)
+
+    def Close(self):
+        self.__db.close()
